+++ cmdb-core/src/main/java/com/webank/cmdb/controller/ApiV2Controller.java
@@ -512,31 +512,4 @@ public class ApiV2Controller {
         staticDtoService.delete(RoleCiTypeCtrlAttrConditionDto.class, requestIds);
     }
     
-	@GetMapping("/model/export")
-    public QueryResponse<CiTypeDto> exportModel(HttpServletResponse response) {
-    	return wrapperService.exportModel();
-    }
-    
-    @DeleteMapping("/model/init")
-    public void initModel() {
-    	wrapperService.initModel();
-    }
-    
-    @PostMapping("/model/showDifferences")
-    public ResponseDto<Map<String,List<CiTypeDto>>> showDifferences(@RequestBody MultipartFile file) {
-    	QueryResponse<CiTypeDto> currentModel = wrapperService.exportModel();
-    	List<CiTypeDto> importModel = wrapperService.importModel(file);
-    	Map<String,List<CiTypeDto>> modelData = new HashMap<String, List<CiTypeDto>>();
-    	modelData.put("currentModel", currentModel.getContents());
-    	modelData.put("importModel", importModel);
-    	ResponseDto<Map<String,List<CiTypeDto>>> responseDto = new ResponseDto<>();
-    	responseDto.setStatusCode("200");
-    	responseDto.setData(modelData);
-    	return responseDto;
-    }
-    
-    @PostMapping("/model/apply")
-    public void applyModel(@RequestBody List<CiTypeDto> ciModel) {
-    	wrapperService.applyModel(ciModel);
-    }
 }
